Fix clippy warnings

+++ nautilus_core/core/src/cvec.rs
@@ -91,7 +91,7 @@ mod tests {
     #[test]
     #[allow(unused_assignments)]
     fn access_values_test() {
-        let test_data = vec![1 as u64, 2, 3];
+        let test_data = vec![1_u64, 2, 3];
         let mut vec_len = 0;
         let mut vec_cap = 0;
         let cvec: CVec = {
